♻️ ShelterMarker를 재사용하기 위해 컴포넌트 및 api 파일분리
